see prev

+++ emulator.cpp
@@ -2391,7 +2391,7 @@ int main(int argn, char ** args)
 	// This should get us the value of something interesting.
 	// Should give us 3c1c0043 in unsigned decimal
 	printf("Lover of the russian queen. %u\n", electricrock->debugGetValue(0x004010e0, 0));
-	electricrock->runEmulation(bv->GetEntryPoint());
+	electricrock->runEmulation((uint32_t)bv->GetEntryPoint());
 
 	// Proper shutdown of core
 	BNShutdown();
